+++ example/lib/main.dart
@@ -1,12 +1,12 @@
 import 'package:enhanced_meteorify/enhanced_meteorify.dart';
-import 'package:flutter/material.dart';
 import 'package:example/app/app_module.dart';
+import 'package:flutter/material.dart';
 import 'package:flutter_modular/flutter_modular.dart';
 
 void main() async {
   try {
     WidgetsFlutterBinding.ensureInitialized();
-    await Meteor.connect('ws://10.0.2.2:3000/websocket');
+    await Meteor.connect('ws://192.168.88.28:3000/websocket');
     runApp(ModularApp(module: AppModule()));
   } on MeteorError catch (error) {
     print(error);