+++ lib/enhanced_meteorify.dart
@@ -3,6 +3,5 @@
 /// Connect your web or flutter apps, written in Dart, with the Meteor framework.
 library enhanced_meteorify;
 
-export 'src/accounts/accounts.dart';
 export 'src/meteor/meteor.dart';
 export 'src/meteor/subscribed_collection.dart';